Fix sorted addition of new, largest orig-src
If the new session entry we want to add has the largest address + message
id hash then hlist_for_each_entry() will iterate over all entries.
However, after it finishes the session pointer won't be the last one
from the list, but instead will be NULL. Which in turn will lead to
adding the new session entry to the list head instead of the tail.
The messed up order can then lead to adding duplicate entries, too.

To fix this and to add an entry to the tail return the last/tail
element instead, which is still in the prev pointer.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>