Fix taking other SAP tools on the same interface into account
We can have more than one SAP tool running on the same interface, for
instance to announce other SDP messages. Even though the SDP messages
are different they might still map to the samp SAP group.

This means that an SAP packet we are receiving with our own IP source
address might not necessarily be from our own SAP tool instance. So
instead of ignoring it we should take them into account for the
bandwidth limit, too.

The only way we currently have to detect and ignore our own packets is
by checking the IP source, the orig-source and the message ID hash all
together.

Note: For now, we also ignore packets with the same orig-source and
message ID hash and a different IP source. We shouldn't detect + add
them as independent sessions. However we should, in the future,
detect them as high-availability SAP announcements and reduce our own
interval accordingly.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>