Increase interval accuracy with low intervals and bw-limit hits
When hitting the bandwidth limitter we would determine an interval in
seconds. This can be a bit inaccurate when we are also using a fast
interval setting.

So to gain some more accuracy here calculate the bandwidth limit in
milliseconds instead (and as a trade-off reduce the overall bandwidth
limit setting maximum by that).

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>