Add more control options for SAP destination address
VLC for instance allows to set/overwrite the SAP destination address
with a custom multicast address.

Even if they might not be RFC conformant add them for compatibility and
debugging reasons.

This now allows to use "-S" for a direct SAP destination address,
without the payload destination to SAP destination mapping. And also a
"-D" option was added to disable autodiscovering payload destinations
(for SAP destinations) from an SDP content.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>