Add blake2 based, deterministic message ID hash (+ as default)
Having a deterministic message ID from the SAP (SDP) payload makes it
easier to use multiple libsap / sap tool instances on different hosts in
a high-availability setup. Also/therefore make this the default, if
the libb2 library is available for hashing.

An option to set a random message ID again was also added to the library
and sap tool.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>